+++ azure/cosmos/__init__.py
@@ -1,1 +1,2 @@
-from .client import Client as CosmosClient
+from .client import Client as CosmosClient, Container, Database, DatabaseResponse
+from internal.cosmos.errors import HTTPFailure
